requirements.txt
streamlit==1.55.0
youtube-transcript-api==1.2.4
google-genai==1.68.0
